requirements.txt
streamlit
google-generativeai
langgraph
pandas
python-dotenv
pytest
faiss-cpu
pypdf


